requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
pyyaml
paho-mqtt
urllib3
pydantic
pymysql
jupyter
requests